using byte

+++ source/com/BombingGames/WurfelEngine/Core/Map/Map.java
@@ -133,12 +133,12 @@ public class Map implements Cloneable {
      * @param generator the custom generator
      */
     public void fill(Generator generator){
-        int chunkpos = 0;
+        byte chunkpos = 0;
         for (byte y=-1; y < 2; y++)
             for (byte x=-1; x < 2; x++){
                 coordlist[chunkpos][0] = x;
                 coordlist[chunkpos][1] = y;  
-                insertChunk((byte) chunkpos, new Chunk(filename, x, y, generator));
+                insertChunk(chunkpos, new Chunk(filename, x, y, generator));
                 chunkpos++;
            }
     }
